backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.24
sqlalchemy[asyncio]>=2.0
alembic>=1.12
pydantic>=2.5
pandas>=2.1
openpyxl>=3.1
python-dotenv>=1.0
psycopg[binary,pool]>=3.1
